requirements.txt
Django==5.0.7
djangorestframework==3.15.2
python-decouple==3.8
celery==5.4.0
redis==5.0.7
requests==2.32.3
django-cors-headers==4.3.1
